requirements.txt
selenium==4.25.0
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.8.3
undetected-chromedriver==3.5.5
setuptools==80.9.0